backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pandas>=1.5.0
openai>=0.27.0
hdbscan>=0.8.0
scikit-learn>=1.2.0
python-dotenv>=0.21.0
numpy>=1.23.0
openpyxl>=3.0.9
